requirements.txt
asn1crypto==0.22.0
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
cryptography==2.0.3
enum34==1.1.6
Flask==0.12.2
idna==2.5
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
pycparser==2.18
PyJWT==1.5.2
pyOpenSSL==17.2.0
python-mpd2==0.5.5
pytz==2017.2
requests==2.18.3
six==1.10.0
twilio==6.5.1
urllib3==1.22
Werkzeug==0.12.2
